本を削除するとき著者も削除するように修正

+++ app/api/books/[id]/route.ts
@@ -6,21 +6,51 @@ export async function DELETE(req: Request, { params }: { params: { id: string } 
   const prisma = new PrismaClient();
   const { id } = params; 
   try {
-    // TODO: authorがDBに残ってしまう
-    const deletedBook = await prisma.book.delete({
+    const bookWithAuthors = await prisma.book.findUnique({
       where: { id },
+      include: {
+        authors: true,
+      },
     });
 
-    return new Response(JSON.stringify({ message: `Book with ID ${id} deleted successfully.`, deletedBook }), {
-      status: 200,
-      headers: { 'Content-Type': 'application/json' },
-    });
+    if (bookWithAuthors) {
+      for (const author of bookWithAuthors.authors) {
+        await prisma.author.delete({
+          where: { id: author.id },
+        });
+      }
+
+      const deletedBook = await prisma.book.delete({
+        where: { id },
+      });
+
+      return new Response(
+        JSON.stringify({
+          message: `Book with ID ${id} and its authors deleted successfully.`,
+          deletedBook,
+        }),
+        {
+          status: 200,
+          headers: { 'Content-Type': 'application/json' },
+        }
+      );
+    } else {
+      return new Response(
+        JSON.stringify({ error: `Book with ID ${id} not found.` }),
+        {
+          status: 404,
+          headers: { 'Content-Type': 'application/json' },
+        }
+      );
+    }
   } catch (error) {
-    console.error(error)
-    return new Response(JSON.stringify({ error: 'Error deleting book' }), {
-      status: 500,
-      headers: { 'Content-Type': 'application/json' },
-    });
+    return new Response(
+      JSON.stringify({ error: 'Error deleting book and authors' }),
+      {
+        status: 500,
+        headers: { 'Content-Type': 'application/json' },
+      }
+    );
   }
   finally
   {
